+++ src/convexHull/visualOptions.js
@@ -94,9 +94,9 @@ export const visualOptions = {
     requiredDimensions: ['x', 'y'],
   },
 
-  pointsRadius: {
+  dotsDiameter: {
     type: 'number',
-    label: 'Dots radius',
+    label: 'Dots diameter',
     default: 2,
     group: 'chart',
     disabled: {