+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## upcoming
 
 - Fixed dimensions and visual options names
+- all the charts now uses `diameter` as visual variable instead of `radius`
 
 ## v1.0.0-beta.18